Extracted shared behaviour between countable and summable into a separate cacheable trait
