+++ src/Behaviours/Countable.php
@@ -1,10 +1,10 @@
 <?php
 namespace Eloquence\Behaviours;
 
-use Illuminate\Support\Facades\DB;
-
 trait Countable
 {
+    use Cacheable;
+
     /**
      * Boot the countable behaviour and setup the appropriate event bindings.
      */
@@ -12,7 +12,7 @@ trait Countable
     {
         static::created(function($model) {
             $model->applyToCountCache(function($config) use ($model) {
-                $model->updateCountCache($config, '+', $model->{$config['foreignKey']});
+                $model->updateCacheRecord($config, '+', 1, $model->{$config['foreignKey']});
             });
         });
 
@@ -22,7 +22,7 @@ trait Countable
 
         static::deleted(function($model) {
             $model->applyToCountCache(function($config) use ($model) {
-                $model->updateCountCache($config, '-', $model->{$config['foreignKey']});
+                $model->updateCacheRecord($config, '-', 1, $model->{$config['foreignKey']});
             });
         });
     }
@@ -48,37 +48,10 @@ trait Countable
             $foreignKey = $this->key($config['foreignKey']);
 
             if ($this->getOriginal($foreignKey) && $this->{$foreignKey} != $this->getOriginal($foreignKey)) {
-                $this->updateCountCache($config, '-', $this->getOriginal($foreignKey));
-                $this->updateCountCache($config, '+', $this->{$foreignKey});
+                $this->updateCacheRecord($config, '-', 1, $this->getOriginal($foreignKey));
+                $this->updateCacheRecord($config, '+', 1, $this->{$foreignKey});
             }
         });
-    }
-
-    /**
-     * Updates a table's record based on the query information provided in the $config variable.
-     *
-     * @param array $config
-     * @param string $operation Whether to increment or decrement a value. Valid values: +/-
-     * @param string $foreignKey
-     */
-    protected function updateCountCache(array $config, $operation, $foreignKey)
-    {
-        if (is_null($foreignKey)) {
-            return;
-        }
-
-        $table = $this->getModelTable($config['model']);
-
-        // the following is required for camel-cased models, in case users are defining their attributes as camelCase
-        $field = snake_case($config['field']);
-        $key = snake_case($config['key']);
-        $foreignKey = snake_case($foreignKey);
-
-        // Execute as a single query as this is more performant and works atomically across both MyISAM and
-        // transactional database engines, resulting in a consistent result.
-        $sql = "UPDATE `{$table}` SET `{$field}` = `{$field}` {$operation} 1 WHERE `{$key}` = {$foreignKey}";
-
-        return DB::update($sql);
     }
 
     /**
@@ -143,76 +116,7 @@ trait Countable
     }
 
     /**
-     * Returns the table for a given model. Model can be an Eloquent model object, or a full namespaced
-     * class string.
-     *
-     * @param string|object $model
-     * @return mixed
-     */
-    public function getModelTable($model)
-    {
-        if (! is_object($model)) {
-            $model = new $model;
-        }
-
-        return $model->getTable();
-    }
-
-    /**
-     * Creates the key based on model properties and rules.
-     *
-     * @param $model
-     * @param string $field
-     * @return string
-     */
-    protected function field($model, $field)
-    {
-        $class = strtolower(class_basename($model));
-        $field = $class.'_'.$field;
-
-        return $field;
-    }
-
-    /**
-     * Returns the true key for a given field.
-     *
-     * @param string $field
-     * @return mixed
-     */
-    protected function key($field)
-    {
-        if (method_exists($this, 'getTrueKey')) {
-            return $this->getTrueKey($field);
-        }
-
-        return $field;
-    }
-
-    /**
-     * Should return an array of the count caches that need to be updated when this
-     * model's state changes. Use the following array below as an example when a User
-     * needs to update a Role's user count cache. These represent the default values used
-     * by the behaviour.
-     *
-     *   return [ 'user_count' => [ 'Role', 'role_id', 'id' ] ];
-     *
-     * So, to extend, the first argument should be an index representing the counter cache
-     * field on the associated model. Next is a numerical array:
-     *
-     * 0 = The model to be used for the update
-     * 1 = The foreign_key for the relationship that RelatedCount will watch *optional
-     * 2 = The remote field that represents the key *optional
-     *
-     * If the latter 2 options are not provided, or if the counter cache option is a string representing
-     * the model, then RelatedCount will assume the ID fields based on conventional standards.
-     *
-     * Ie. another way to setup a counter cache is like below. This is an identical configuration to above.
-     *
-     *   return [ 'user_count' => 'Role' ];
-     *
-     * This can be simplified even further, like this:
-     *
-     *   return [ 'Role' ];
+     * Return the count cache configuration for the model.
      *
      * @return array
      */
