+++ src/Database/Model.php
@@ -13,6 +13,6 @@ use Eloquence\Behaviours\Uuid;
  */
 abstract class Model extends \Illuminate\Database\Eloquent\Model
 {
-    use \Eloquence\Behaviours\CamelCasing;
-    use \Eloquence\Behaviours\Uuid;
+    use CamelCasing;
+    use Uuid;
 }
